+++ components/micropython/port/src/omv/py/py_mjpeg.c
@@ -25,7 +25,7 @@ typedef struct py_mjpeg_obj {
     mp_obj_t fp;
 } py_mjpeg_obj_t;
 
-static mp_obj_t py_mjpeg_open(uint n_args, const mp_obj_t *args, mp_map_t *kw_args)
+static mp_obj_t py_mjpeg_open(size_t n_args, const mp_obj_t *args, mp_map_t *kw_args)
 {
     py_mjpeg_obj_t *mjpeg = m_new_obj(py_mjpeg_obj_t);
     mjpeg->width  = py_helper_keyword_int(n_args, args, 1, kw_args, MP_OBJ_NEW_QSTR(MP_QSTR_width), MAIN_FB()->w);
@@ -57,7 +57,7 @@ static mp_obj_t py_mjpeg_size(mp_obj_t mjpeg_obj)
     return mp_obj_new_int(file_size(arg_mjpeg->fp));
 }
 
-static mp_obj_t py_mjpeg_add_frame(uint n_args, const mp_obj_t *args, mp_map_t *kw_args)
+static mp_obj_t py_mjpeg_add_frame(size_t n_args, const mp_obj_t *args, mp_map_t *kw_args)
 {
     py_mjpeg_obj_t *arg_mjpeg = args[0];
     image_t *arg_img = py_image_cobj(args[1]);
